Feature/finalizacao controllers

+++ MedicalCare/Controllers/DietaController.cs
@@ -1,8 +1,11 @@
 ﻿using MedicalCare.DTO;
+using MedicalCare.Enums;
 using MedicalCare.Interfaces;
+using MedicalCare.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
+using System.Security.Cryptography;
 
 
 namespace MedicalCare.Controllers
@@ -14,12 +17,14 @@ namespace MedicalCare.Controllers
         private readonly IDietaService _dietaService;
         private readonly IPacienteService _pacienteService;
         private readonly IUsuarioService _usuarioService;
+        private readonly ILogService _logService;
 
-        public DietaController(IDietaService dietaService, IPacienteService pacienteService, IUsuarioService usuarioService)
+        public DietaController(IDietaService dietaService, IPacienteService pacienteService, IUsuarioService usuarioService, ILogService logService)
         {
             _dietaService = dietaService;
             _pacienteService = pacienteService;
             _usuarioService = usuarioService;
+            _logService = logService;
         }
 
         [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
@@ -28,14 +33,36 @@ namespace MedicalCare.Controllers
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+
                 var verificaSeExsitePaciente = _pacienteService.GetById(dietaCreateDto.PacienteId);
                 var verificaSeExisteUsuario = _usuarioService.GetById(dietaCreateDto.UsuarioId);
-                if (verificaSeExsitePaciente != null && verificaSeExisteUsuario != null)
+                if (verificaSeExsitePaciente == null || verificaSeExisteUsuario == null)
                 {
-                DietaGetDto dietaGet = _dietaService.CreateDieta(dietaCreateDto);
-                return Created("Dieta salva com sucesso.", dietaGet);
+                    return NoContent();
                 }
-                return NoContent();
+
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                if (tipo == "Médico" || tipo =="Enfermeiro")
+                {
+                    dietaCreateDto.UsuarioId = _id;
+                }
+                DietaGetDto dietaGet = _dietaService.CreateDieta(dietaCreateDto);
+
+                LogModel logModel = new LogModel
+                {
+                    Descricao = $"{tipo} {nome}, de Id {_id}, cadastrou a dieta de id {dietaGet.Id}.",
+                    Dominio = "Dieta-cadastro."
+                };
+                _logService.CreateLog(logModel);
+
+                return Created("Dieta salva com sucesso.", dietaGet);
             }
             catch (Exception)
             {
@@ -49,14 +76,39 @@ namespace MedicalCare.Controllers
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                int id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+
                 if (pacienteId.HasValue)
                 {
                     var dietas = _dietaService.GetDietasByPaciente(pacienteId.Value);
+
+                    LogModel logModel = new LogModel
+                    {
+                        Descricao = $"{tipo} {nome}, de Id {id}, listou dietas do paciente de id {pacienteId}.",
+                        Dominio = "Dieta-obter."
+                    };
+                    _logService.CreateLog(logModel);
+
                     return Ok(dietas);
                 }
                 else
                 {
                     var dietas = _dietaService.GetAllDietas();
+
+                    LogModel logModel = new LogModel
+                    {
+                        Descricao = $"{tipo} {nome}, de Id {id}, listou todas as dietas.",
+                        Dominio = "Dieta-obter."
+                    };
+                    _logService.CreateLog(logModel);
+
                     return Ok(dietas);
                 }
             }
@@ -65,23 +117,6 @@ namespace MedicalCare.Controllers
                 return StatusCode(HttpStatusCode.InternalServerError.GetHashCode(), "Erro interno.");
             }
         }
-        //[HttpGet("{id}")]
-        //public ActionResult<DietaGetDto> Get([FromRoute] int id)
-        //{
-        //    try
-        //    {
-        //        DietaGetDto dietaGet = _dietaService.GetById(id);
-        //        if (dietaGet == null)
-        //        {
-        //            return NotFound("Id de dieta não encontrado");
-        //        }
-        //        return Ok(dietaGet);
-        //    }
-        //    catch (Exception)
-        //    {
-        //        return StatusCode(HttpStatusCode.InternalServerError.GetHashCode(), "Erro interno.");
-        //    }
-        //}
 
         [Authorize(Roles = "Administrador, Médico, Enfermeiro")]
         [HttpPut("{id}")]
@@ -89,12 +124,42 @@ namespace MedicalCare.Controllers
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+
                 DietaGetDto verificaSeExiste = _dietaService.GetById(id);
                 if (verificaSeExiste == null)
                 {
                     return NoContent();
                 }
+
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+                if (tipo == "Médico" || tipo =="Enfermeiro")
+                {
+                    dietaUpdateDto.UsuarioId = _id;
+                }
+
+                var verificaSeExsitePaciente = _pacienteService.GetById(dietaUpdateDto.PacienteId);
+                var verificaSeExisteUsuario = _usuarioService.GetById(dietaUpdateDto.UsuarioId);
+                if (verificaSeExisteUsuario == null || verificaSeExsitePaciente == null)
+                {
+                    return NoContent();
+                }
+
                 DietaGetDto dietaGet = _dietaService.UpdateDieta(dietaUpdateDto, id);
+
+                LogModel logModel = new LogModel
+                {
+                    Descricao = $"{tipo} {nome}, de Id {_id}, atualizou a dieta de id {id}.",
+                    Dominio = "Dieta-atualizar."
+                };
+                _logService.CreateLog(logModel);
+
                 return Ok(dietaGet);
             }
             catch (Exception)
@@ -109,9 +174,24 @@ namespace MedicalCare.Controllers
         {
             try
             {
+                var ativo = bool.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "StatusDoSistema").Value);
+                if (!ativo)
+                {
+                    return BadRequest("Usuário inativo no sistema");
+                }
+                int _id = int.Parse(HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Id").Value);
+                var nome = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Nome").Value;
+                var tipo = HttpContext.User.Claims.FirstOrDefault(x => x.Type == "Tipo").Value;
+
                 bool remocao = _dietaService.DeleteDieta(id);
                 if (remocao)
                 {
+                    LogModel logModel = new LogModel
+                    {
+                        Descricao = $"{tipo} {nome}, de Id {_id}, excluiu a dieta de id {id}.",
+                        Dominio = "Dieta-excluir."
+                    };
+                    _logService.CreateLog(logModel);
                     return Accepted();
                 }
                 return NoContent();
